# api/openapi.yaml
openapi: 3.1.0
info:
  title: Jewelcase API
  version: 1.0.0-draft
  summary: The public API of a Jewelcase server, used by first-party and third-party clients alike.
  description: |
    The contract between Jewelcase servers and every client. First-party clients use
    nothing outside it (`design/general.md` §6). Requirements cited as
    `requirements/file.md` §N.

    ## Conventions

    **Base path.** Paths are relative to `{basePath}/api/v1`. A server may sit under any
    prefix behind a reverse proxy; clients never assume the domain root.

    **Authentication.** Login registers a device and returns an opaque token. Send it as
    `Authorization: Bearer <token>`. The web client may instead rely on the
    `jewelcase_session` cookie set at login; cookie-authenticated requests that change
    state must carry an `X-Jewelcase-Client` header or they are refused. Any authenticated
    endpoint may answer `401` — see `Problem.reason` for whether a device keeps its
    downloads (`requirements/offline.md` §9).

    **Scope.** Paths under `/libraries/{libraryId}` are scoped to that library. Anything
    the caller cannot reach answers `404`, exactly as if it did not exist; an action the
    caller's role forbids answers `403` (`requirements/users.md` §10). Identifiers in a
    request select within the caller's scope and never widen it.

    **Personal data.** `/me` is the caller's own account. Admins read other accounts'
...

...

...
    never on a collection. Lists report `total`.

...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /me/plugins:
    get:
      tags: [Account]
      summary: List the plugins the caller can connect
      description: |
        Plugins that act for each person, such as a scrobbler, connected with the person's own
        account on its service (`requirements/plugins.md` §4.1, §6): every installed plugin
        with personal settings that is enabled in a library the caller reaches.
      operationId: listMyPlugins
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/PersonalPlugin' }
        default: { $ref: '#/components/responses/Problem' }
...
  /me/plugins/{pluginId}/search-sharing:
    parameters:
      - $ref: '#/components/parameters/PluginId'
    put:
      tags: [Account]
      summary: Turn sharing the caller's searches with a plugin on or off
      description: |
        Off until the caller turns it on (`requirements/search.md` §6). On, the plugin is sent
        each search the caller settles on from then, never one before. Off, or disconnecting,
        asks it to forget every search it was sent; what it already sent on is not recalled
        (`requirements/users.md` §7). Turning it on needs the caller to have connected the
        plugin, and the plugin to ask for searches (`asksForSearches`), or answers `422`.
      operationId: setMySearchSharing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SearchSharing' }
      responses:
        '204': { description: Set. }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        default: { $ref: '#/components/responses/Problem' }

  # ───────────────────────────── Devices ─────────────────────────────
...
  /libraries/{libraryId}/search/{section}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - name: section
        in: path
        required: true
        schema: { $ref: '#/components/schemas/SearchSectionType' }
    get:
      tags: [Search]
      summary: Expand a search section
      description: The full list for one section, in match order or sorted (`requirements/search.md` §4).
      operationId: searchSection
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Within'
        - $ref: '#/components/parameters/Genre'
        - $ref: '#/components/parameters/YearFrom'
        - $ref: '#/components/parameters/YearTo'
        - $ref: '#/components/parameters/Played'
        - $ref: '#/components/parameters/PlayedBefore'
        - $ref: '#/components/parameters/PlayedAfter'
        - $ref: '#/components/parameters/MinPlayCount'
        - name: sort
          in: query
          description: '`relevance`, or any sort the section''s browse endpoint accepts.'
          schema: { type: string, default: relevance }
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResultPage' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
...
  /libraries/{libraryId}/recent-searches/{recentSearchId}:
    parameters:
      - $ref: '#/components/parameters/LibraryId'
      - name: recentSearchId
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Search]
      summary: Remove a recent search
      operationId: deleteRecentSearch
      responses:
        '204': { description: Removed. }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }

  # ───────────────────────────── Sync ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque per-device token from login.
    cookieAuth:
      type: apiKey
      in: cookie
      name: jewelcase_session
      description: Set at login for browser clients. State-changing requests must also send `X-Jewelcase-Client`.
  parameters:
    LibraryId:
      name: libraryId
      in: path
      required: true
      schema: { type: string }
...
    PluginId:
      name: pluginId
      in: path
      required: true
      schema: { type: string }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
...
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    Genre:
      name: genre
      in: query
      description: Tag IDs of type `genre`.
      style: form
      explode: false
      schema:
        type: array
        items: { type: string }
...
    YearFrom:
      name: yearFrom
      in: query
      schema: { type: integer }
    YearTo:
      name: yearTo
      in: query
      schema: { type: integer }
...
    Played:
      name: played
      in: query
      description: Filter by whether the caller has ever played the item.
      schema:
        type: string
        enum: [never, ever]
    PlayedBefore:
      name: lastPlayedBefore
      in: query
      description: Last played before this time — "not played since".
      schema: { type: string, format: date-time }
    PlayedAfter:
      name: lastPlayedAfter
      in: query
      schema: { type: string, format: date-time }
    MinPlayCount:
      name: minPlayCount
      in: query
      schema: { type: integer, minimum: 0 }
...
    Query:
      name: q
      in: query
      required: true
      description: Search text, from the first character typed.
      schema: { type: string, minLength: 1 }
...
    Within:
      name: within
      in: query
      description: 'Narrow to a container, as `artist:ID`, `album:ID`, or `playlist:ID` (`requirements/search.md` §5).'
      schema: { type: string, pattern: '^(artist|album|playlist):.+$' }
...
  responses:
    Problem:
      description: Error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: '`not_found` — nonexistent, or out of the caller''s reach.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationFailed:
      description: The request was understood but not acceptable; see `code` and `errors`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail:
          type: string
          description: Human-readable, safe to show. Never contains credentials or plugin internals.
        code:
          type: string
          enum:
            - unauthenticated
            - session_revoked
            - invalid_credentials
            - rate_limited
            - forbidden
            - not_found
            - method_not_allowed
            - validation_failed
            - weak_password
            - breached_password
            - plugin_invalid
            - plugin_exists
            - permissions_required
            - username_taken
            - version_conflict
            - capacity_exceeded
            - cursor_expired
            - setup_required
            - unsupported_media
            - root_overlap
            - root_unavailable
            - device_unavailable
            - not_active_device
            - source_unavailable
            - source_timeout
            - plugin_settings_invalid
            - internal
        reason:
          type: string
          enum: [credentials_changed, logged_out]
          description: |
            With `session_revoked`. `credentials_changed` — the device keeps its downloads and
            logs in again with its `deviceId`; `logged_out` — the device removes them
            (`requirements/offline.md` §9).
        errors:
          type: array
          description: Field-level problems, with `validation_failed` and `plugin_settings_invalid`.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string, description: JSON Pointer into the request body. }
              message: { type: string }
        currentVersion:
          type: integer
          description: With `version_conflict`.
    PageInfo:
      type: object
      required: [nextCursor, total]
      properties:
        nextCursor: { type: [string, 'null'] }
        total: { type: integer }
    Id:
      type: string
      description: Opaque, stable identifier.
...
    PartialDate:
      type: string
      pattern: '^\d{4}(-\d{2}(-\d{2})?)?$'
      description: A date at tag precision.
    DurationUs:
      type: integer
      format: int64
      minimum: 0
      description: Microseconds.
    Availability:
      type: string
      enum: [available, missing]
...
    ImageRef:
      type: object
      required: [id, placeholder]
      properties:
        id:
          type: string
          description: Fetch from `getImage`. Changes whenever the image does.
        placeholder:
          type: string
          x-status: draft
          description: A few bytes, base64, decoding to a recognizable correctly-coloured stand-in (`requirements/offline.md` §1.1).
    ArtistCredit:
      type: object
      required: [id, name]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        name:
          type: [string, 'null']
          description: '`null` for the unknown artist.'
    Personal:
      type: object
      description: Personal data on an entity (`requirements/conventions.md` §1). Never shared between accounts.
      required: [playCount, lastPlayedAt]
      properties:
        playCount: { type: integer }
        lastPlayedAt: { type: [string, 'null'], format: date-time }
...
    AlbumRef:
      type: object
      required: [id, title, artists, image]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        title:
          type: [string, 'null']
          description: '`null` for the unknown album.'
        artists:
          type: array
          items: { $ref: '#/components/schemas/ArtistCredit' }
        image:
          description: The album's one cover, shown for every track on it (`requirements/albums.md` §5).
          oneOf:
            - $ref: '#/components/schemas/ImageRef'
            - type: 'null'
    TagRef:
      type: object
      required: [id, name]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        name: { type: string }
    TrackSummary:
      type: object
      required: [id, title, artists, album, durationUs, availability, audio, explicit]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        title:
          type: string
          description: Falls back to the filename for untagged files.
        artists:
          type: array
          description: Ordered; the first is primary (`requirements/conventions.md` §7).
          items: { $ref: '#/components/schemas/ArtistCredit' }
        album: { $ref: '#/components/schemas/AlbumRef' }
        discNumber: { type: [integer, 'null'] }
        trackNumber: { type: [integer, 'null'] }
        durationUs: { $ref: '#/components/schemas/DurationUs' }
        releaseDate:
          oneOf:
            - $ref: '#/components/schemas/PartialDate'
            - type: 'null'
        genres:
          type: array
          items: { $ref: '#/components/schemas/TagRef' }
        explicit: { type: boolean }
        audio:
          description: Enough to tell a FLAC from an MP3 at a glance (`requirements/tracks.md` §4). The full `Track` extends it.
          allOf:
            - $ref: '#/components/schemas/AudioSummary'
        availability: { $ref: '#/components/schemas/Availability' }
        addedAt: { type: string, format: date-time }
        personal:
          description: The caller's own listening — the viewed account's under `/admin/users/{userId}`. Absent from catalog parts and the library feed.
          allOf:
            - $ref: '#/components/schemas/Personal'
    AudioSummary:
      type: object
      required: [codec, lossless]
      properties:
        codec: { type: string }
        lossless: { type: boolean }
...
    AlbumType:
      type: string
      enum: [album, ep, single, compilation]
      description: '`requirements/albums.md` §3.'
    AlbumSummary:
      type: object
      required: [id, title, artists, type, trackCount, trackTotal, discCount, discTotal, durationUs, image, availability]
      description: |
        Completeness has two parts (`requirements/albums.md` §4): tracks missing from discs you
        have (`trackCount` below `trackTotal`), and discs you do not have (`discCount` below
        `discTotal`). Incomplete is information, never an error.
      properties:
        id: { $ref: '#/components/schemas/Id' }
        title: { type: [string, 'null'] }
        artists:
          type: array
          items: { $ref: '#/components/schemas/ArtistCredit' }
        releaseDate:
          oneOf:
            - $ref: '#/components/schemas/PartialDate'
            - type: 'null'
        type: { $ref: '#/components/schemas/AlbumType' }
        genres:
          type: array
          items: { $ref: '#/components/schemas/TagRef' }
        trackCount:
          type: integer
          description: |
            Tracks from this album in the library, including any marked missing — those stay in
            position with their own `availability`, so "incomplete" means never had, not lost.
        trackTotal:
          type: [integer, 'null']
          description: |
            Sum of the tagged per-disc totals for the discs present. Track totals are tagged per
            disc, so an absent disc contributes nothing; see `discCount`. `null` when no disc
            present has a tagged total.
        discCount:
          type: integer
          description: Discs with at least one track in the library.
        discTotal:
          type: [integer, 'null']
          description: Tagged total number of discs.
        durationUs: { $ref: '#/components/schemas/DurationUs' }
        image:
          oneOf:
            - $ref: '#/components/schemas/ImageRef'
            - type: 'null'
        availability: { $ref: '#/components/schemas/Availability' }
        addedAt: { type: string, format: date-time }
        personal:
          description: The caller's own listening — the viewed account's under `/admin/users/{userId}`. Absent from catalog parts and the library feed.
          allOf:
            - $ref: '#/components/schemas/Personal'
...
    ArtistSummary:
      type: object
      required: [id, name, image, albumCount, trackCount]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        name: { type: [string, 'null'] }
        image:
          oneOf:
            - $ref: '#/components/schemas/ImageRef'
            - type: 'null'
        albumCount:
          type: integer
          description: Albums the artist owns.
        trackCount:
          type: integer
          description: Owned and featured tracks.
        addedAt: { type: string, format: date-time }
        personal:
          description: The caller's own listening — the viewed account's under `/admin/users/{userId}`. Absent from catalog parts and the library feed.
          allOf:
            - $ref: '#/components/schemas/Personal'
...
    SearchSectionType:
      type: string
      enum: [tracks, albums, artists, playlists, lyrics]
    LyricMatch:
      type: object
      required: [track, line, before, after]
      properties:
        track: { $ref: '#/components/schemas/TrackSummary' }
        line:
          type: object
          required: [text]
          properties:
            text: { type: string }
            startMs: { type: [integer, 'null'] }
        before:
          type: array
          description: Lines of context before the match.
          items: { type: string }
        after:
          type: array
          items: { type: string }
    SearchResult:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [track, album, artist, playlist, lyric]
        track: { $ref: '#/components/schemas/TrackSummary' }
        album: { $ref: '#/components/schemas/AlbumSummary' }
        artist: { $ref: '#/components/schemas/ArtistSummary' }
        playlist: { $ref: '#/components/schemas/Playlist' }
        lyric: { $ref: '#/components/schemas/LyricMatch' }
...
    SearchResultPage:
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/SearchResult' }
...
    Playlist:
      type: object
      required: [id, libraryId, title, description, image, trackCount, durationUs, pinned, createdAt, modifiedAt, version]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        libraryId: { $ref: '#/components/schemas/Id' }
        title: { type: string }
        description: { type: [string, 'null'] }
        image:
          description: Custom, or a mosaic of the tracks inside; `null` for an empty playlist.
          oneOf:
            - $ref: '#/components/schemas/ImageRef'
            - type: 'null'
        customImage: { type: boolean }
        trackCount: { type: integer }
        missingCount: { type: integer }
        durationUs: { $ref: '#/components/schemas/DurationUs' }
        pinned: { type: boolean }
        createdAt: { type: string, format: date-time }
        modifiedAt: { type: string, format: date-time }
        version:
          type: integer
          description: Advances with every change; edits apply only against the version they were made on.
        personal:
          description: The caller's own listening — the viewed account's under `/admin/users/{userId}`. Absent from catalog parts and the library feed.
          allOf:
            - $ref: '#/components/schemas/Personal'
...
    PersonalPlugin:
      type: object
      required: [id, name, connected, asksForSearches, sharesSearches]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        connected:
          type: boolean
          description: Whether the caller has saved settings for it, and so it acts for them.
        asksForSearches:
          type: boolean
          description: |
            Whether the caller could share their searches with it: it asks for them, and an admin
            allowed it.
        sharesSearches:
          type: boolean
          description: Whether the caller has turned on sharing their searches with it.
    SearchSharing:
      type: object
      required: [sharing]
      properties:
        sharing: { type: boolean }
...